+++ src/audio_config.h
@@ -6,6 +6,3 @@ void AudioConfigInitialize(HMODULE module);
 void AudioConfigReload();
 bool AudioConfigIsEnabled();
 void AudioConfigSetEnabled(bool enabled);
-bool AudioConfigHotkeyEnabled();
-int AudioConfigHotkeyKey();
-int AudioConfigHotkeyModifier();